+++ src/js/components/slider.js
@@ -1,6 +1,7 @@
 import Slider, { speedUp } from '../mixin/slider';
 import SliderReactive from '../mixin/internal/slider-reactive';
 import TransitionerPlugin from './internal/slider-transitioner';
+import ParallaxPlugin from './internal/slider-parallax';
 
 function plugin(UIkit) {
 
@@ -11,8 +12,10 @@ function plugin(UIkit) {
     UIkit.use(Slider);
 
     var {mixin} = UIkit;
-    var {$, $$, css, data, includes, isNumeric, toggleClass, toFloat} = UIkit.util;
+    var {$, $$, addClass, css, data, includes, isNumeric, isUndefined, toggleClass, toFloat} = UIkit.util;
     var Transitioner = TransitionerPlugin(UIkit);
+
+    UIkit.component('slider-parallax', ParallaxPlugin(UIkit, 'slider'));
 
     UIkit.component('slider', {
 
@@ -144,10 +147,11 @@ function plugin(UIkit) {
                     this.index = this.getValidIndex();
                 }
 
-                var diff = Math.abs(this.index + (this.dir > 0
-                        ? this.index < this.prevIndex ? this.maxIndex + 1 : 0
-                        : this.index > this.prevIndex ? -this.maxIndex : 0
-                ) - this.prevIndex);
+                var diff = Math.abs(
+                    this.index
+                    - this.prevIndex
+                    + (this.dir > 0 && this.index < this.prevIndex || this.dir < 0 && this.index > this.prevIndex ? (this.maxIndex + 1) * this.dir : 0)
+                );
 
                 if (!this.dragging && diff > 1) {
 
@@ -168,6 +172,16 @@ function plugin(UIkit) {
 
                 this.reorder();
 
+            },
+
+            itemshow() {
+                !isUndefined(this.prevIndex) && addClass(this._getTransitioner().getItemIn(), this.clsActive);
+            },
+
+            itemshown() {
+                var actives = this._getTransitioner(this.index).getActives();
+                this.slides.forEach(slide => toggleClass(slide, this.clsActive, includes(actives, slide)));
+                (!this.sets || includes(this.sets, toFloat(this.index))) && this.slides.forEach(slide => toggleClass(slide, this.clsActivated, includes(actives, slide)));
             }
 
         },
@@ -200,7 +214,7 @@ function plugin(UIkit) {
                     width = this.list.offsetWidth / 2 - next.offsetWidth / 2,
                     j = 0;
 
-                while (width >= 0) {
+                while (width > 0) {
                     var slideIndex = this.getIndex(--j + index, index),
                         slide = this.slides[slideIndex];