+++ src/js/mixin/internal/slider-drag.js
@@ -39,9 +39,16 @@ export default function (UIkit) {
                 },
 
                 handler(e) {
-                    if (isTouch(e) || !hasTextNodesOnly(e.target)) {
-                        this.start(e);
+
+                    if (!isTouch(e) && hasTextNodesOnly(e.target)
+                        || e.button > 0
+                        || this.length < 2
+                        || this.preventCatch
+                    ) {
+                        return;
                     }
+
+                    this.start(e);
                 }
 
             },
@@ -58,15 +65,7 @@ export default function (UIkit) {
 
         methods: {
 
-            start(e) {
-
-                if (e.button > 0 || this.length < 2) {
-                    return;
-                }
-
-                if (this.preventCatch) {
-                    return;
-                }
+            start() {
 
                 this.drag = this.pos;
 
@@ -127,12 +126,13 @@ export default function (UIkit) {
                 var prev = slides[prevIndex],
                     next = slides[nextIndex],
                     changed = this.index !== nextIndex,
-                    edge = prevIndex === nextIndex;
+                    edge = prevIndex === nextIndex,
+                    reset;
 
                 [this.index, this.prevIndex].filter(i => !includes([nextIndex, prevIndex], i)).forEach(i => {
                     trigger(slides[i], 'itemhidden', [this]);
 
-                    this._transitioner && this._transitioner.reset();
+                    reset = true;
 
                     if (edge) {
                         this.prevIndex = prevIndex;
@@ -140,7 +140,7 @@ export default function (UIkit) {
 
                 });
 
-                if (this.index === prevIndex && this.prevIndex !== prevIndex) {
+                if (this.index === prevIndex && this.prevIndex !== prevIndex || reset && edge) {
                     trigger(slides[this.index], 'itemshown', [this]);
                 }
 
@@ -152,6 +152,7 @@ export default function (UIkit) {
                     trigger(next, 'beforeitemshow', [this]);
                 }
 
+                reset && this._transitioner && this._transitioner.reset();
                 this._transitioner = this._translate(Math.abs(this.percent), prev, !edge && next);
 
                 if (changed) {
@@ -206,4 +207,4 @@ export default function (UIkit) {
         return !el.children.length && el.childNodes.length;
     }
 
-};
+}